requirements.txt
numpy>=1.11.3
scipy>=0.19
matplotlib>=1.4.3
Pillow>=2.7.0
lxml>=3.5.0
scikit-image>=0.14.0
opencv-python-headless
